Fix all clippy warnings across workspace
Resolve all clippy warnings to achieve a clean clippy pass:

- Remove unnecessary parentheses (hll.rs)
- Use slice iteration instead of index loops (clustering.rs, thread_ref.rs)
- Rename ExpansionStrategy::from_str to parse to avoid std::str::FromStr confusion
- Use div_ceil instead of manual (n+1)/2 (faker.rs)
- Remove unnecessary same-type cast (faker.rs)
- Flatten nested if-let in read_dir loop (init.rs)
- Simplify redundant closures, map_or, and identity maps (auto-fixed)
- Add #[allow(clippy::too_many_arguments)] for internal builder functions
  where refactoring would be invasive with no external API benefit

+++ crates/knit-learn/src/streaming/hll.rs
@@ -67,7 +67,7 @@ impl HyperLogLog {
         let remaining = hash >> self.precision;
         // Count leading zeros in the remaining bits + 1 (ρ function).
         // For a k-bit word of all zeros, ρ = k + 1.
-        let bits = (64 - self.precision) as u8;
+        let bits = 64 - self.precision;
         let rho = if remaining == 0 {
             bits + 1
         } else {
